agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "financial-analyst-agent"
version = "0.2.0"
description = "Autonomous financial analyst — monitors a multi-brokerage portfolio via Plaid and delivers Claude-powered buy/sell/hold recommendations"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"

dependencies = [
    # Plaid — brokerage connectivity
    "plaid-python>=25.0.0",
    "python-dotenv>=1.0.0",

    # Jupyter — notebook pipeline runner
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "nbconvert>=7.0.0",

    # Data
    "pandas>=2.0.0",

    # Market data & technicals
    "yfinance>=0.2.30",
    "pandas-ta>=0.3.14",
    "requests>=2.31.0",

    # Claude API
    "anthropic>=0.40.0",

    # Notifications
    "sendgrid>=6.10.0",

    # Supabase sync
    "supabase>=2.28.0",

    # Account setup server
    "flask>=3.0.0",

    # Token encryption
    "cryptography>=42.0.0",
]

[project.scripts]
fin-pipeline = "finanalyst.cli:pipeline"
fin-sync     = "finanalyst.cli:sync"
fin-poller   = "finanalyst.cli:poller"
fin-server   = "finanalyst.cli:server"

[tool.hatch.build.targets.wheel]
packages = ["finanalyst"]

[tool.hatch.build.targets.sdist]
include = [
    "finanalyst/",
    "notebooks/",
    "*.py",
    "*.md",
    "*.toml",
    ".env.example",
    "scheduler/",
    "static/",
]
